requirements.txt
google-cloud-pubsub==1.0.2
google-cloud-iot==2.6.4
PyMySQL==0.9.3
sqlalchemy==1.3.8
